feat: add basetag
